Extract quota goal boundary projection helpers
